+++ lib/src/internal.dart
@@ -1,0 +1,4 @@
+import 'package:meta/meta.dart';
+
+@internal
+T identity<T>(T t) => t;